Rewriting Chapter 2 Finger Print cache, use slab allocator to manage memory, use linked list | binary search tree | skip list as different search algorithm.

+++ c_cpp/0002/build.sh
@@ -15,15 +15,20 @@ fi
 
 echo "mydef=$mydef"
 
-rm -f *.o
-rm -f libcommon*
-gcc -D$mydef -fPIC -c -o common.o common.c
-gcc -D$mydef -shared -fPIC -o libcommon.so common.o
-#main.c
-gcc -D$mydef -Wl,-rpath=./ -o mytest mytest.c -L.  -lcommon -lcrypto
+
+if [ $mydef == LINKEDLIST ]
+then
+    cd linkedlist
+    rm -f common.o
+    rm -f libcommon*
+
+    gcc -D$mydef -fPIC -c -o common.o common.c
+    gcc -D$mydef -shared -fPIC -o libcommon.so common.o
+    
+    #main.c
+    cd ../
+    gcc -D$mydef -Wl,-rpath=./ -o mytest mytest.c -L./linkedlist  -lcommon -lcrypto
+    echo "Complile mytest $mydef success"
+fi
 
 
-#cd ptmalloc2
-#make shared
-
-echo "Complile mytest success"
